refactor: table 초기 생성 로직 위치 변경
- 각 모듈에서 mysql Module로 일원화
- cors 초기 설정
