+++ backend/src/domain/category/category.service.ts
@@ -1,25 +1,13 @@
-import { Injectable, OnModuleInit } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { MySQLService } from 'src/config/mysql/mysql.service';
 import format from '../utils/format';
 import { categoryCreateType, categoryUpdateType } from './category.type';
 
 @Injectable()
-export class CategoryService implements OnModuleInit {
+export class CategoryService {
   promisePool: any;
   constructor(private mysqlService: MySQLService) {
     this.promisePool = this.mysqlService.pool.promise();
-  }
-  async onModuleInit() {
-    const poolPromise = this.mysqlService.pool.promise();
-    await poolPromise.execute(`
-      CREATE TABLE IF NOT EXISTS CATEGORY (
-        id INT PRIMARY KEY AUTO_INCREMENT,
-        name VARCHAR(20) NOT NULL,
-        storeId INT,
-        deletedAt DATETIME,
-        FOREIGN KEY (storeId) REFERENCES STORE(id)
-      )
-    `);
   }
 
   async getCategoryByStoreId(storeId: number) {
