+++ backend/src/domain/receipt/receiptToProduct/receiptToProduct.service.ts
@@ -1,26 +1,13 @@
-import { Injectable, OnModuleInit } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { MySQLService } from 'src/config/mysql/mysql.service';
 import format from 'src/domain/utils/format';
 import { rTpCreateType } from './receiptToProduct.type';
 
 @Injectable()
-export class ReceiptToProductService implements OnModuleInit {
+export class ReceiptToProductService {
   promisePool: any;
   constructor(private mysqlService: MySQLService) {
     this.promisePool = this.mysqlService.pool.promise();
-  }
-
-  async onModuleInit() {
-    await this.promisePool.execute(`
-    CREATE TABLE IF NOT EXISTS RECEIPT_TO_PRODUCT (
-      receiptId INT,
-      productId INT,
-      count INT,
-      FOREIGN KEY (receiptId) REFERENCES RECEIPT(id),
-      FOREIGN KEY (productId) REFERENCES PRODUCT(id),
-      PRIMARY KEY (receiptId, productId)
-    )
-  `);
   }
 
   async createReceipToProduct(rTpData: rTpCreateType) {
