+++ backend/src/domain/store/store.type.ts
@@ -1,4 +1,5 @@
 export interface storeCreateType {
+  storeId: string;
   name: string;
   password: string;
   branchName?: string;
