+++ backend/src/main.ts
@@ -12,6 +12,7 @@ async function bootstrap() {
   app.setGlobalPrefix('api');
   app.use(cookieParser());
 
+  app.enableCors();
   await app.listen(port);
 }
 bootstrap();
